requirements.txt
streamlit>=1.35.0
gspread>=6.1.2
google-auth>=2.29.0
pandas>=2.2.2
plotly>=5.22.0
numpy>=1.26.4
